src/azure-video-indexer/requirements.txt
azure-core
azure-functions
azure-search-documents
requests